added error for double accept

+++ app/controllers/requests.rb
@@ -26,17 +26,24 @@ class MakersBnb < Sinatra::Base
     days.each do |y|
       filter_dates << y.date
     end
-    filter_dates.each do |day|
-      remove_day = Property.get(property_id).days(:conditions => {:date => day})
-      remove_day.destroy!
-      Property.get(property_id).days(:conditions => {:date => day}).each do |x|
-        Day.get(x.id).destroy!
+    if (Property.get(property_id).days & days) == days
+      filter_dates.each do |day|
+        remove_day = Property.get(property_id).days(:conditions => {:date => day})
+        remove_day.destroy!
+        Property.get(property_id).days(:conditions => {:date => day}).each do |x|
+          Day.get(x.id).destroy!
+        end
       end
-
+      Request.get(request_id).destroy!
+      redirect '/requests'
+    else
+      flash.now[:errors] = "Property has already been booked for these dates. Please decline."
+      @made_requests = Property.all.requests(booker_id: current_user.id)
+      user_properties = Property.all(user_id: current_user.id)
+      @received_requests = user_properties.all.requests
+      erb :'requests/index'
     end
-    Request.get(request_id).destroy!
-    redirect '/requests'
-  end
+end
 
   post '/decline' do
     request_id = params[:request]
